requirement.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coreapi==2.3.3
coreschema==0.0.4
distro==1.9.0
Django==4.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jiter==0.12.0
legacy-cgi==2.6.4
MarkupSafe==3.0.3
openai==2.7.1
openapi-codec==1.3.2
packaging==25.0
prompt_toolkit==3.0.52
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
setuptools==80.9.0
simplejson==3.20.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
